Dialog editor extracted in a dedicated file + some doc
